have the icon call baz's github as the base url. return that url so that it can just the src attribute value
when i try to add the relative, local path as the src, it doesn't work. it's also too much work to individually put the svg elements (in xml format) into the json as there are too many options. this option relies heavily on the icon still being hosted at that url by basmilius. but, it does save having to keep those svgs in the repo, saving up space